feat: added pagination
As the backend roadmap suggest for improve is pagination, so I added
one. Need some improvements, but I could work on that after studying
more python's magic.

Usage:
n for next page
p for previous page
q for quit/exit

it will "clear" the texts on the terminal when navigating between pages